requirements.txt
Flask
Flask-Login 
passlib
Flask-WTF 
FLask-Migrate 
Flask-SQLAlchemy 
pyotp 
qrcode 
python-decouple
python-dotenv
Pillow
bleach
zxcvbn
numpy
cryptography
flask-limiter
pymemcache
uwsgi
flask_paranoid
pycryptodome
wtforms[email]
argon2_cffi
pyclamd
ipaddress